+++ CHANGELOG.md
@@ -23,3 +23,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - automatically register for all alarm types
   - create Docker image
   - create Debian package
+  - support `climate/remoteClimateState/set` with `off`, `on` and `front`